+++ src/main/java/frc/robot/constants/TestRobotConstants.java
@@ -13,7 +13,7 @@ import edu.wpi.first.units.Angle;
 import edu.wpi.first.units.Distance;
 import edu.wpi.first.units.Measure;
 import edu.wpi.first.units.Velocity;
-import frc.robot.subsystems.base.DriveBase.WheelModuleIndex;
+import frc.robot.subsystems.swerve_module.IndexedSwerveModule.WheelModuleIndex;
 
 public class TestRobotConstants extends AbstractConstants {
